不要なsliceを削除しました
