--- src/components/codeArea/defaultButtonCss.ts
@@ -1,17 +1,0 @@
-import { arrayType } from '../../types/cssTypes'
-
-export const defaultButtonCss: arrayType = {
-    color: 'rgba(255,0,0,1)',
-    backgroundColor: 'rgba(0,0,255,1)',
-    border: 'none',
-    padding: '0px',
-    textDecoration: 'none',
-    display: 'inline-block',
-    fontSize: '14px',
-    borderColor: 'none',
-    borderStyle: 'none',
-    borderWidth: '0px',
-    borderRadius: '0px',
-    width: '200px',
-    height: '100px',
-}
