+++ src/store.ts
@@ -1,15 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit'
-import buttonViewSlice from './components/buttonView/buttonViewSlice'
 import cssCustomAnimeAreaSlice from './components/cssCustomAnimeArea/cssCustomAnimeAreaSlice'
-// import cssCustomAreaSlice from './components/cssCustomArea/cssCustomAreaSlice'
 import pseudoAreaSlice from './components/pseudoArea/pseudoAreaSlice'
 import counterReducer from './redux_test/counterSlice'
 
 export const store = configureStore({
     reducer: {
         counter: counterReducer,
-        buttonView: buttonViewSlice,
-        // cssCustomArea: cssCustomAreaSlice,
         cssCustomAnimeArea: cssCustomAnimeAreaSlice,
         pseudoArea: pseudoAreaSlice,
     },
